Update collaborative-filter.ipynb
MovieLens dataset is one of the most popular benchmark datasets used to test the potency of 
various collaborative filtering models. The full version of the dataset consists of more than 
26,000,000 ratings, 45,000 movies by 270,000 users. However, for the sake of fast execution, 
we will be using a smaller dataset. To do so, from https://www.kaggle.com/prajitdatta/movielens-100k-dataset/version/1, download the dataset, 
which is a snapshot of MovieLens dataset that contains 100,000 ratings applied by 1,000 users to 1,700 movies. Your task is to build a user-based collaborative filtering based on the user ratings provided in the dataset.